Remove workarounds for border radius in (RN)
So it looks like RN now supports this for the `Image` component! There looks to be issues border width and colour (they don't show up at all), but that's on RN's side.

I think this isn't a breaking change. Technically, you go from `{ borderRadius: 50 }` to `{ borderTopLeftRadius: 50, borderTopRightRadius: … }`, but this shouldn't actually change much

If we remove the blacklist here, we'll remove it in css-to-react-native and there'll be a tiny perf win